+++ src/main/java/com/github/ddth/cacheadapter/KryoCacheEntrySerializer.java
@@ -1,40 +1,33 @@
 package com.github.ddth.cacheadapter;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-
-import com.esotericsoftware.kryo.Kryo;
-import com.esotericsoftware.kryo.io.Input;
-import com.esotericsoftware.kryo.io.Output;
-import com.esotericsoftware.kryo.pool.KryoCallback;
-import com.esotericsoftware.kryo.pool.KryoFactory;
-import com.esotericsoftware.kryo.pool.KryoPool;
+import com.github.ddth.cacheadapter.utils.KryoUtils;
 
 /**
- * This implementation of {@link ICacheEntrySerializer} use Kryo library
+ * This implementation of {@link ICacheEntrySerializer} uses Kryo library
  * (https://github.com/EsotericSoftware/kryo) to serialize/deserialize cache
- * entries..
+ * entries.
  * 
  * @author Thanh Nguyen <btnguyen2k@gmail.com>
  * @since 0.3.0
  */
-public class KryoCacheEntrySerializer implements ICacheEntrySerializer {
+public class KryoCacheEntrySerializer extends AbstractCacheEntrySerializer {
 
-    private KryoPool kryoPool;
-
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public KryoCacheEntrySerializer init() {
-        KryoFactory factory = new KryoFactory() {
-            public Kryo create() {
-                Kryo kryo = new Kryo();
-                return kryo;
-            }
-        };
-        kryoPool = new KryoPool.Builder(factory).softReferences().build();
+        super.init();
 
         return this;
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public void destroy() {
+        super.destroy();
     }
 
     /**
@@ -42,20 +35,7 @@ public class KryoCacheEntrySerializer implements ICacheEntrySerializer {
      */
     @Override
     public byte[] serialize(final CacheEntry ce) {
-        if (ce == null) {
-            return null;
-        }
-        return kryoPool.run(new KryoCallback<byte[]>() {
-            @Override
-            public byte[] execute(Kryo kryo) {
-                ByteArrayOutputStream baos = new ByteArrayOutputStream();
-                Output output = new Output(baos);
-                kryo.writeObject(output, ce);
-                output.flush();
-                output.close();
-                return baos.toByteArray();
-            }
-        });
+        return KryoUtils.serialize(ce);
     }
 
     /**
@@ -63,16 +43,7 @@ public class KryoCacheEntrySerializer implements ICacheEntrySerializer {
      */
     @Override
     public CacheEntry deserialize(final byte[] data) {
-        if (data == null) {
-            return null;
-        }
-        return kryoPool.run(new KryoCallback<CacheEntry>() {
-            @Override
-            public CacheEntry execute(Kryo kryo) {
-                Input input = new Input(new ByteArrayInputStream(data));
-                return kryo.readObject(input, CacheEntry.class);
-            }
-        });
+        return KryoUtils.deserialize(data, CacheEntry.class);
     }
 
 }